Cleans up pseudocode in main_loop

+++ micromouse_main/src/util/conversions.cpp
@@ -3,7 +3,7 @@
 #include "conversions.h"
 
 /* Converts encoder ticks to a distance in mm */
-int ticksToMM(int ticks) {
+double ticksToMM(int ticks) {
     // doTheThing();
     return 0;
 }
